requirements.txt
flask==3.0.0
requests==2.31.0
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
werkzeug==3.0.1
matplotlib==3.10.7